# apps/api/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Convixx API",
    "description": "Multi-tenant AI Voice Calling Platform backend. RAG-based question answering with self-hosted LLM and OpenAI fallback.\n\n**Documentation:** Interactive UI at `/docs`. Raw OpenAPI JSON at `/docs/json` (same origin as this API).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host (recommended for Try it out)"
    },
    {
      "url": "https://convixx.in",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Admin token for customer APIs. Set in .env as ADMIN_TOKEN."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Customer API key. Scopes KB, agents, ask, chat, voice to that customer."
      }
    },
    "schemas": {
      "SarvamSettingsResponse": {
        "type": "object",
        "properties": {
          "stt_provider": {
            "type": "string"
          },
          "stt_model": {
            "type": "string"
          },
          "tts_provider": {
            "type": "string"
          },
          "tts_model": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/settings/rag": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get RAG LLM mode",
        "description": "Returns `rag_use_openai_only`: when true, RAG uses OpenAI only (self-hosted LLM is not called). Stored per customer. Requires x-api-key.",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current setting",
            "content": {
              "application/json": {
                "example": {
                  "rag_use_openai_only": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Set RAG LLM mode (OpenAI-only toggle)",
        "description": "Set `rag_use_openai_only`. When true, `/ask` and `/ask/voice` use OpenAI for RAG answers only (same KB context and prompts; self-hosted is skipped). When false, default pipeline (self-hosted with OpenAI fallback) applies. Requires x-api-key.",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rag_use_openai_only"
                ],
                "properties": {
                  "rag_use_openai_only": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated setting",
            "content": {
              "application/json": {
                "example": {
                  "rag_use_openai_only": true
                }
              }
            }
          }
        }
      }
    },
    "/settings/sarvam": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get Sarvam settings",
        "description": "Returns current Sarvam-related STT and TTS settings. Requires x-api-key.",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current Sarvam settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SarvamSettingsResponse"
                },
                "example": {
                  "stt_provider": "sarvam",
                  "stt_model": "saaras:v3",
                  "tts_provider": "sarvam",
                  "tts_model": "bulbul:v2"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get full customer settings",
        "description": "Returns the full `customer_settings` row for the authenticated customer. Defaults are guaranteed — every customer has a row (auto-created on customer insert, or backfilled by migration 005).",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full customer settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSettings"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Customer not found"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Update customer settings (tenant scope)",
        "description": "Partial update of `customer_settings`. Admin-only fields (see `CustomerSettingsPatch` description) are rejected with 403. Send only the fields you want to change — omitted fields are left untouched.",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerSettingsPatch"
              },
              "example": {
                "voicebot_multilingual": true,
                "default_language_code": "hi-IN",
                "allowed_language_codes": [
                  "hi-IN",
                  "en-IN",
                  "mr-IN"
                ],
                "rag_use_openai_only": true,
                "rag_top_k": 4,
                "barge_in_enabled": true,
                "barge_in_mode": "immediate",
                "allow_language_switch": true,
                "stop_words": [
                  "stop",
                  "operator",
                  "main menu"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSettings"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "formErrors": [],
                    "fieldErrors": {
                      "rag_top_k": [
                        "Expected number, received string"
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "One or more fields are admin-only",
            "content": {
              "application/json": {
                "example": {
                  "error": "Some fields are admin-only",
                  "admin_only_fields": [
                    "tts_model",
                    "vad_silence_timeout_ms"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/customers/{customerId}/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Admin: get customer settings",
        "description": "Admin view of `customer_settings` for any customer. Requires `x-admin-token`.",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full customer settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSettings"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Customer not found"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Admin: update any customer settings field",
        "description": "Partial update of `customer_settings` with no scope check — admins may edit every field, including admin-only ones.",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerSettingsPatch"
              },
              "example": {
                "voicebot_enabled": true,
                "tts_model": "bulbul:v3",
                "tts_default_sample_rate": 8000,
                "vad_silence_timeout_ms": 1200,
                "max_utterance_buffer_bytes": 10485760,
                "max_concurrent_calls": 50,
                "webhook_secret": "whsec_XXXXXXXXXXXX",
                "webhook_retry_attempts": 5
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSettings"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
    "/admin/customers/{customerId}/settings/reset": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Admin: reset settings to defaults",
        "description": "Deletes the customer's `customer_settings` row and re-inserts a fresh one, so every column reverts to its database default (see `CustomerSettings` schema).",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fresh (default) settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSettings"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
...
    "/settings/industry-context": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get industry context",
        "description": "Free-form JSON used to steer RAG tone/domain grounding for this tenant.",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "industry_context": {}
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Customer not found"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Update industry context",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "industry_context"
                ],
                "properties": {
                  "industry_context": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "example": {
                  "industry_context": {}
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Validation error"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Customer not found"
                }
              }
            }
          }
        }
      }
    },
...
  }
}